Create AdminActivity for admins to manage other users.
Updated MainActivity "Administration" button to go to the AdminActivity. Created spinner in AdminActivity that shows current users in database. Created delete button and wired up functionality to UserDAO and UserRepository so that it will delete a user from the table.

Added buttons to clear user's shopping list and kitchen but needs implementation still.